requirements.txt
azure-iot-device~=2.0.0
azure-storage-blob
